+++ laktory/models/pipeline/orchestrators/lakeflowdeclarativepipelineorchestrator.py
@@ -141,7 +141,7 @@ class LakeflowDeclarativePipelineOrchestrator(Pipeline, PipelineChild):
         excludes["type"] = True
         excludes["dataframe_backend"] = True
         excludes["dataframe_api"] = True
-        excludes["full_refresh_mode"] = True
+        excludes["reset_mode"] = True
         return excludes
 
     @property